airflow/requirements.txt
pandas
apache-airflow-providers-apache-spark
deezer-python
spotipy
requests
pywebhdfs